+++ feed/index.html
@@ -16,7 +16,7 @@
     <script src="/js/feedposts.js" type="module"></script>
     <script src="/js/searchPosts.js" type="module"></script>
     <script src="/js/filterFeedPosts.js" type="module"></script>
-    <script src="/js/backToTopButton.js" type="module"></script>
+    <script src="/js/utils/backToTopButton.js" type="module"></script>
     <script src="/js/createPost.js" type="module"></script>
     <script src="/js/populateFeed.js" type="module"></script>
   </head>
